+++ DTCD-EventSystem/src/utils/CustomAction.js
@@ -1,11 +1,16 @@
 export class CustomAction {
-  constructor(actionName, guid, callback, args = null) {
+  static generateID(guid, actionName) {
+    return `${guid}[${actionName}]`;
+  }
+
+  constructor(guid, actionName, callback, args = null) {
     this.guid = guid;
     this.name = actionName;
     this.args = args;
-    this.id = `${guid}[${actionName}]`;
+    this.id = CustomAction.generateID(guid, actionName);
     this.callback = callback;
   }
+
   toString() {
     return this.id;
   }