+++ CHANGELOG.md
@@ -5,8 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2021-02-11
+## [0.3.0]
+
+### Added
+
+- integrated log system
+- token to subscription
+- custom actions support
+- added proper event args comparison
+- removeCustomAction method
+
 ### Changed
+
+- changed version of SDK
+- all logs now recorded with 'debug' level
+
+## Fixed
+
+- renamed log method to a new one
+
+## [0.2.0] - 2021-02-11
+
+### Changed
+
 - code source directory name to DTCD-EventSystem
 - downloading DTCD-SDK from [storage](http://storage.dev.isgneuro.com)
 - paths in source files to DTCD-SDK